适配新裁判系统串口协议，提高了机器人交互数据发送频率到30Hz

+++ Module/Referee/referee_robot_interaction_manager.c
@@ -55,8 +55,8 @@ uint8_t *refereeEncodeRobotInteractionData(robot_interaction_type_e robot_intera
 {
     uint8_t *sub_data = NULL;
 
-    if ((*manager.step_clock - manager.last_successful_send_time) >
-        (ROBOT_INTERACTION_COMM_TIME_MS / manager.step_time))
+    if ((float)(*manager.step_clock - manager.last_successful_send_time) >
+        (ROBOT_INTERACTION_COMM_TIME_MS / (float)manager.step_time))
     {
         switch (robot_interaction_type)
         {
